+++ test/cli.test.js
@@ -20,6 +20,7 @@ import CLI from '../src/cli.js';
 describe('CLI Test', () => {
   afterEach(() => {
     delete process.env.HLX_TEST;
+    delete process.env.CUSTOM_ENV_VAR;
   });
 
   it('has correct defaults with no arguments', () => {
@@ -31,7 +32,6 @@ describe('CLI Test', () => {
     assert.equal(builder.cfg.build, true);
     assert.equal(builder.cfg.minify, false);
     assert.equal(builder.cfg.esm, false);
-    assert.equal(builder.cfg.bundler, 'webpack');
     assert.equal(builder.cfg.test, undefined);
     assert.equal(builder.cfg.showHints, true);
     assert.equal(builder.cfg.nodeVersion, '18');
@@ -123,12 +123,6 @@ describe('CLI Test', () => {
     const builder = new CLI()
       .prepare(['--esm']);
     assert.equal(builder.cfg.esm, true);
-  });
-
-  it('sets bundler', () => {
-    const builder = new CLI()
-      .prepare(['--bundler', 'rollup']);
-    assert.equal(builder.cfg.bundler, 'rollup');
   });
 
   it('sets test flag', () => {
@@ -438,6 +432,7 @@ describe('CLI Test', () => {
       apiId: 'someapi',
       arch: 'arm64',
       cleanUpIntegrations: false,
+      cleanUpVersions: false,
       createRoutes: false,
       // eslint-disable-next-line no-template-curly-in-string
       lambdaFormat: '${packageName}--${baseName}',
@@ -447,6 +442,37 @@ describe('CLI Test', () => {
       identitySources: ['$request.header.Authorization'],
       deployBucket: '',
       updateSecrets: undefined,
+      tracingMode: undefined,
+      extraPermissions: undefined,
+      layers: undefined,
+      logFormat: undefined,
+      tags: undefined,
     });
+  });
+
+  it('interpolates environment variables', () => {
+    process.env.CUSTOM_ENV_VAR = 'test';
+    const builder = new CLI()
+      .prepare([
+        '--test',
+        // eslint-disable-next-line no-template-curly-in-string
+        'some-${env.CUSTOM_ENV_VAR}-value',
+        '--serverless-externals',
+        // eslint-disable-next-line no-template-curly-in-string
+        '/${env.CUSTOM_ENV_VAR}/index.html',
+        '--serverless-externals',
+        // eslint-disable-next-line no-template-curly-in-string
+        '/${env.CUSTOM_ENV_VAR}/index.txt',
+      ]);
+    assert.deepEqual(builder.cfg.test, 'some-test-value');
+    assert.deepEqual(builder.cfg.serverlessExternals, ['/test/index.html', '/test/index.txt']);
+  });
+
+  it('ignores non-existing environment variables', () => {
+    const builder = new CLI()
+      // eslint-disable-next-line no-template-curly-in-string
+      .prepare(['--test', 'some-${env.CUSTOM_ENV_VAR}-value']);
+    // eslint-disable-next-line no-template-curly-in-string
+    assert.deepEqual(builder.cfg.test, 'some-${env.CUSTOM_ENV_VAR}-value');
   });
 });